fixed hw-06
