--- helpers/isAccesibleFile.js
@@ -1,12 +1,0 @@
-const fs = require("fs/promises");
-const isAccessible = (path) => {
-  fs.access(path)
-    .then(() => true)
-    .catch(() => false);
-};
-const createFolderIsNotExist = async (folder) => {
-  if (!(await isAccessible(folder))) {
-    await fs.mkdir(folder);
-  }
-};
-module.exports = createFolderIsNotExist;
