+++ helpers/sendEmail.js
@@ -1,0 +1,20 @@
+const sgMail = require("@sendgrid/mail");
+require("dotenv").config();
+const { SENDGRID_API_KEY } = process.env;
+sgMail.setApiKey(SENDGRID_API_KEY);
+const sendEmail = async (data) => {
+  const msg = {
+    from: "zagrousaspear@gmail.com",
+    ...data,
+  };
+  await sgMail
+    .send(msg)
+    .then(() => {
+      console.log("Email sent");
+    })
+    .catch((error) => {
+      console.error(error);
+    });
+  return true;
+};
+module.exports = sendEmail;
